requirements.txt
uvicorn==0.21.1
fastapi==0.109.1
starlette==0.35.0
pydantic==2.9.2
databases==0.7.0
SQLAlchemy==1.4.47
aiomysql
httpx==0.23.0
redis
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
openai==0.28
